db/requirements.txt
alembic
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt
